web/backend/requirements/requirements.txt
Django==1.11.4
django-appconf==1.0.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.4
django-filter==2.0.0
django-imagekit==4.0.2
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
Jinja2==2.10
MarkupSafe==1.1.0
pilkit==2.0
pytz==2018.7
six==1.12.0
sqlparse==0.2.4

# Raven
raven==6.5.0
